+++ engine/src/main/java/net/jqwik/engine/support/JqwikKotlinSupport.java
@@ -1,37 +1,42 @@
-package net.jqwik.engine.discovery;
+package net.jqwik.engine.support;
 
 import java.lang.annotation.*;
 import java.lang.reflect.*;
 
 public class JqwikKotlinSupport {
 
-	public static boolean isInternalKotlinMethod(Method method) {
-		Class<?> aClass = method.getDeclaringClass();
-		return isKotlinClass(aClass) && isKotlinInternal(method);
-	}
-
-	public static boolean isSpeciallyNamedKotlinMethod(Method method) {
-		Class<?> aClass = method.getDeclaringClass();
-		return isKotlinClass(aClass) && isKotlinSpecial(method);
+	public static String javaOrKotlinName(Method targetMethod) {
+		String name = targetMethod.getName();
+		if (isKotlinClass(targetMethod.getDeclaringClass())) {
+			if (isKotlinInternal(targetMethod)) {
+				name = nameWithoutInternalPart(name);
+			}
+			if (isKotlinSpecial(targetMethod)) {
+				name = nameWithoutSpecialPart(name);
+			}
+		}
+		return name;
 	}
 
 	private static boolean isKotlinInternal(Method method) {
-		// Kotlin appends a module extension to internal method names
-		// The exact text of this extension is not known (to me)
-		return method.getName().lastIndexOf('$') > 0;
+		if ((method.getModifiers() & Modifier.PUBLIC) == 0) {
+			return false;
+		}
+		// Kotlin appends "$kotlin" to internal method names in Java bytecode
+		return method.getName().endsWith("$kotlin");
 	}
 
 	private static boolean isKotlinSpecial(Method method) {
-		// Kotlin appends an 7-char-extension separated by a hyphen to method names in special cases
+		// Kotlin appends a 7-char-extension separated by a hyphen to method names in special cases
 		// e.g. when method has UInt parameter, the original method is appended with '-WZ4Q5Ns'
 		// TODO: Find out what's really happening here
-		String name = method.getName();
+		String name = isKotlinInternal(method) ? nameWithoutInternalPart(method.getName()): method.getName();
 		int lastIndexOfHyphen = name.lastIndexOf('-');
 		return lastIndexOfHyphen >= 0 && lastIndexOfHyphen == (name.length() - 8);
 	}
 
 	// Learned mechanism to detect Kotlin class in https://stackoverflow.com/a/39806722
-	public static boolean isKotlinClass(Class<?> aClass) {
+	private static boolean isKotlinClass(Class<?> aClass) {
 		for (Annotation annotation : aClass.getDeclaredAnnotations()) {
 			if (annotation.annotationType().getTypeName().equals("kotlin.Metadata")) {
 				return true;
@@ -40,14 +45,14 @@ public class JqwikKotlinSupport {
 		return false;
 	}
 
-	public static String nameWithoutInternalPart(Method targetMethod) {
-		int lastDollarPosition = targetMethod.getName().lastIndexOf('$');
-		return targetMethod.getName().substring(0, lastDollarPosition);
+	private static String nameWithoutInternalPart(String name) {
+		int lastDollarPosition = name.lastIndexOf('$');
+		return name.substring(0, lastDollarPosition);
 	}
 
-	public static String nameWithoutSpecialPart(Method targetMethod) {
-		int lastDollarPosition = targetMethod.getName().lastIndexOf('-');
-		return targetMethod.getName().substring(0, lastDollarPosition);
+	private static String nameWithoutSpecialPart(String name) {
+		int lastDollarPosition = name.lastIndexOf('-');
+		return name.substring(0, lastDollarPosition);
 	}
 
 	// Kotlin constructor parameters can have default values.